WIP: remove old service integrate into new service
