--- src/main/java/se/isselab/HAnS/featureExtension/backgroundTask/.feature-to-file
@@ -1,17 +1,0 @@
-FeatureFileMappingsBackground.java
-FeatureFileMapping
-
-FeatureMetricsBackground.java
-Scattering, Tangling
-
-FileMappingBackground.java
-FeatureFileMapping
-
-ScatteringDegreeBackground.java
-Scattering
-
-TanglingDegreeBackground.java
-Tangling
-
-TanglingMapBackground.java
-Tangling
