--- src/main/java/se/isselab/HAnS/featureExtension/backgroundTask/.feature-to-folder
@@ -1,1 +1,0 @@
-Callback
