+++ src/main/java/se/isselab/HAnS/featureLocation/FeatureLocationManager.java
@@ -27,7 +27,7 @@ import com.intellij.util.Query;
 import org.jetbrains.annotations.NotNull;
 import se.isselab.HAnS.FeatureAnnotationSearchScope;
 import se.isselab.HAnS.codeAnnotation.psi.*;
-import se.isselab.HAnS.featureExtension.backgroundTask.BackgroundTask;
+import se.isselab.HAnS.pluginExtensions.backgroundTask.BackgroundTask;
 import se.isselab.HAnS.featureModel.FeatureModelUtil;
 import se.isselab.HAnS.featureModel.psi.FeatureModelFeature;
