+++ src/main/java/se/isselab/HAnS/metrics/ProjectMetrics.java
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package se.isselab.HAnS.metrics.v2;
+package se.isselab.HAnS.metrics;
 
 import com.intellij.openapi.util.Pair;
 import se.isselab.HAnS.featureLocation.FeatureFileMapping;
