+++ src/main/java/se/isselab/HAnS/metrics/calculators/FeatureScattering.java
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package se.isselab.HAnS.metrics;
+package se.isselab.HAnS.metrics.calculators;
 
 import com.intellij.openapi.project.Project;
 import se.isselab.HAnS.featureLocation.FeatureFileMapping;
